[PUBDEV-3616] Move multiple AstMerge-related helper files to an appropriate directory
