+++ h2o-core/src/main/java/water/fvec/Frame.java
@@ -6,7 +6,7 @@ import water.*;
 import water.api.schemas3.KeyV3;
 import water.exceptions.H2OIllegalArgumentException;
 import water.parser.BufferedString;
-import water.rapids.Merge;
+import water.rapids.ast.prims.mungers.merge.Merge;
 import water.util.*;
 
 import java.io.IOException;
