+++ h2o-core/src/main/java/water/rapids/ast/prims/mungers/AstMerge.java
@@ -4,8 +4,7 @@ import water.*;
 import water.fvec.*;
 import water.parser.BufferedString;
 import water.rapids.Env;
-import water.rapids.Merge;
-import water.rapids.Val;
+import water.rapids.ast.prims.mungers.merge.Merge;
 import water.rapids.ast.AstRoot;
 import water.rapids.vals.ValFrame;
 import water.rapids.ast.AstPrimitive;
