+++ h2o-core/src/main/java/water/rapids/ast/prims/mungers/merge/RadixCount.java
@@ -1,4 +1,4 @@
-package water.rapids;
+package water.rapids.ast.prims.mungers.merge;
 
 import water.*;
 import water.fvec.Chunk;
