+++ h2o-core/src/test/java/water/rapids/SortTest.java
@@ -5,6 +5,7 @@ import org.junit.Test;
 import water.*;
 import water.fvec.*;
 import water.nbhm.NonBlockingHashMapLong;
+import water.rapids.ast.prims.mungers.merge.Merge;
 import water.rapids.vals.ValFrame;
 
 import java.util.Random;
